# services/sensor-data/src/server/apiSpec/openapi.yaml
openapi: 3.0.3
info:
  title: Sensor Data Service
  description: Changestream POC - Sensor Data Service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
paths:
  "/":
    get:
      summary: Return healthcheck
      description: Return healthcheck
      operationId: returnHealthCheck
      responses:
        '200':
          description: successful
          content:
            plain/text:
              schema:
                "$ref": "#/components/schemas/UptimeInfo"
  "/uptime":
    get:
      summary: Return server uptime
      description: Return server uptime
      operationId: returnServerUptime
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/UptimeInfo"
  "/api-docs/json":
    get:
      summary: Return api docs json
      description: Return api docs json
      operationId: returnApiDocsJSON
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: object
  "/api-docs/html":
    get:
      summary: Return api docs html
      description: Return api docs html
      operationId: returnApiDocsHTML
      responses:
        '200':
          description: successful
          content:
            text/html:
              schema:
                type: string
  "/data":
    put:
      summary: Add new Sensor Data
      description: Add new Sensor Data
      operationId: addSensorData
      requestBody:
        description: New Sensor Data to add
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SensorData"
      responses:
        '204':
          description: Successful
        '400':
          description: Bad request
          content: {}
        '409':
          description: Conflict
          content: {}
    get:
      summary: Query Sensor Data
      description: Returns a Sensor Data document
      operationId: querySensorData
      parameters:
        - in: query
          name: sensorId
          required: true
          schema:
            type: string
          description: The id if the sensor
        - in: query
          name: since
          required: true
          schema:
            type: number
          description: The epoch time since
        - in: query
          name: until
          required: true
          schema:
            type: number
          description: The epoch time until
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SensorData"
        '404':
          description: No Sensor Data found
          content: {}
components:
  schemas:
    SensorData:
      type: object
      required:
        - sensorId
        - time
        - value
      properties:
        sensorId:
          type: string
          format: uuid
        time:
          type: integer
          format: int64
        value:
          type: number
          format: float
    HealthCheck:
      type: string
    UptimeInfo:
      type: object
      required:
        - uptime
      properties:
        uptime:
          type: integer
          format: int64